refactor(native): move invocation accounting into its own module
`WireFormat`, its two function-pointer type aliases, and the accounting
lifecycle around one API invocation — `begin_api_invocation`,
`begin_child_invocation`, `finish_child_invocation`, `api_invocation_snapshot`,
`finish_api_invocation`, `record_accounting_diagnostic`,
`record_context_snapshot`, `context_snapshot_diagnostic`, `wire_format_for` and
their two private helpers — move to `api_process_adapter/invocation.rs`.

`WireFormat`'s eight function-pointer fields become `pub(super)` because
`execution`, `compaction`, and `generation` invoke them. The struct's doc
comment says those pointers stay private, and they still are: `pub(super)`
means visible inside `api_process_adapter`, which is the same boundary the
single file gave them. Bodies are byte-identical.
